+++ src/main/java/it/cnr/ilc/lexo/manager/text/ControlledCommonMarkParser.java
@@ -469,7 +469,6 @@ public final class ControlledCommonMarkParser {
                 || "author".equals(key)
                 || "date".equals(key)
                 || "description".equals(key)
-                || "language".equals(key)
                 || "format".equals(key)
                 || "corpus".equals(key);
     }